Alert disable the window when open
