+++ src/Views/alerts/mod.rs
@@ -4,56 +4,11 @@ use crate::chrono::Datelike;
 use crate::gtk::{Adjustment, Calendar, CheckButton, DropDown, MessageDialog, SpinButton};
 use crate::model::Count::Count;
 
+mod alert;
+mod confirm;
 mod edit;
+
+pub use alert::alert;
+pub use confirm::confirm;
 pub use edit::edit_count;
 static mut HAS_MESSAGE_DIALOG: bool = false;
-
-#[allow(deprecated)]
-pub fn alert(message: &str, title: &str) {
-    unsafe {
-        if !HAS_MESSAGE_DIALOG {
-            let alert_dialog = MessageDialog::builder()
-                .message_type(gtk::MessageType::Warning)
-                .text(message)
-                .title(title)
-                .width_request(350)
-                .height_request(70)
-                .buttons(gtk::ButtonsType::Ok)
-                .build();
-
-            alert_dialog.connect_response(move |alert, _| {
-                alert.destroy();
-                HAS_MESSAGE_DIALOG = false;
-            });
-
-            HAS_MESSAGE_DIALOG = true;
-
-            alert_dialog.show();
-        }
-    }
-}
-
-#[allow(deprecated)]
-pub fn confirm(message: &str, title: &str) -> Option<MessageDialog> {
-    unsafe {
-        if !HAS_MESSAGE_DIALOG {
-            let dialog = MessageDialog::builder()
-                .buttons(gtk::ButtonsType::YesNo)
-                .message_type(gtk::MessageType::Warning)
-                .text(message)
-                .title(title)
-                .build();
-
-            dialog.connect_response(|alert, _| {
-                alert.destroy();
-                HAS_MESSAGE_DIALOG = false;
-            });
-
-            HAS_MESSAGE_DIALOG = true;
-
-            Some(dialog)
-        } else {
-            None
-        }
-    }
-}
